pref: remove unneeded component nesting
